==> DataAnalysis/01_ReadInData.Rmd <==
---
title: "Reading in data"
output:
  pdf_document: default
  html_document: default
date: "2023-03-27"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, results = "hide")
library(RSQLite)
library(tidyverse)
library(magrittr)
library(cowplot)
library(modelr)
```

## Purpose
The purpose of this part is only to read in the data for ribdif/bacdive/growth rate dataset. The data exploration/visualization/analysis will be done in a separate document.
The final data is formatted and saved as a Rdata file for easy use in the further analysis. 


### Read data from ribdif and bacdive from the db 

```{r}
# Connect to the DB
conn <- dbConnect(SQLite(),"../s16.sqlite")


# Read in the data from the database
D_tmp <- dbGetQuery(conn, "SELECT * FROM ribdif_bacdive_joined")
D_tmp <- tibble(D_tmp, .name_repair ="universal")

# set the Antibiotic resistance data types to the correct types
D_tmp <- mutate(D_tmp, across(antibiotics:PH.range, as.character))
## Drop index number originally from the pandas dataframe and drop the NCBI.tax.ID
D_tmp <- select(D_tmp, !c(NCBI.tax.ID,strain.designation))


# Get a few statistics from the raw data before it is joined
# Firstly get number of entries where all 16S genes are the same for comparison with ribogroove
dbGetQuery(conn, "SELECT * FROM species_gcf2species_ribdif_info") %>% 
  filter(number_16s > 1) %>% 
  filter(total_div == 0) %>% 
  nrow

# Then get the median and quantile number of div and 16S gene copy number 
# As the mirror database we used this data to compare to joined at a species level. we did aswell
dbGetQuery(conn, "SELECT * FROM species_gcf2species_ribdif_info") %>% 
  summarise(n16 = mean(number_16s), .by=species..7) %>%
  summarise(median(n16),quantile(n16, 0.75),quantile(n16,0.25))

dbGetQuery(conn, "SELECT * FROM species_gcf2species_ribdif_info") %>% 
  summarise(div = mean(total_div), .by=species..7) %>%
  summarise(median(div),quantile(div, 0.75),quantile(div,0.25))

```
### Read data about growth rates and select the maximum for each

```{r}
# Define function to calculate mode
Mode <- function(x) {
  ux <- unique(x)
  ux[which.max(tabulate(match(x, ux)))]
}

growthD_tmp <- read_csv("./data/biokinetic.csv")
# Rename the df to match the other dataframes
growthD_tmp <- rename(growthD_tmp, species = binomial.name)
# Pick out max value for each species 
growthD <- growthD_tmp %>% 
  group_by(species) %>% 
  summarise(max_growth_min = max(rate.per.minute), 
            aero = Mode(aero),
            trophy = Mode(trophy),
            temp.dataset = mean(T.C) 
            )
  
```
### Read AR data from running abricate

```{r}
# Read in teh data
D_ar <- read_csv("./data/sum_final2.csv")
drug_ann_tmp <- read.csv("./data/megares_annotations_v3.00.csv", sep="|", skip=1, header=FALSE)

# Rename the columns
drug_ann <- drug_ann_tmp %>% 
    select(V3, V2) %>%
    rename(res_type = V3, res_ann = V2)

# Get a list of all of the resistance types which are actuall antibiotic resistances
drug_v <- drug_ann %>% filter(res_ann == "Drugs") %>% pull(res_type) %>% unique()
# Then based on this list select the antibiotic resistance drugs
Ddrug <- D_ar %>% select(gcf,any_of(drug_v))
Dsum <- Ddrug %>% 
  rowwise() %>% 
  mutate(ar_count = sum(pick(2:last_col()))) %>% 
  select(gcf, ar_count)

# set the columns to 0 or 1 as chars
D_is_res <- Dsum %>% mutate(ar_count = ifelse(ar_count == 0, "0", "1"))


# Read data from db
Dar50_tmp <- dbGetQuery(conn, "SELECT * FROM ar_info_full_50")

# Join the read in data with the database db
Dar50_tmp2 <- left_join(Dar50_tmp, D_is_res, by="gcf") 

# If for a genera no antibiotic resistance were found it will be nan in the data, therefore replace NAN with 0
Dar50_tmp3 <- Dar50_tmp2 %>% 
  select(species=species..7, ar_count.y) %>%
  mutate(ar_count.y = replace_na(ar_count.y, "0"))

# And lastly summarise at a species level
Dar50 <- Dar50_tmp3  %>% 
  summarise(ar_count = Mode(ar_count.y), .by = species)


```


### Read data from bacdive gotten directly from it's search function

```{r}

# Read in the data
sporeD_tmp <- read.csv2("./data/spore.csv", sep=",", header = T,skip = 2)
pathoD_tmp <- read.csv2("./data/pathogenicity.csv", sep=",", header = T,skip = 2)
animal_pathoD_tmp <- read.csv2("./data/path_animals.csv", sep=",", header = T,skip = 2)
multicelD_tmp <- read.csv2("./data/multicellular.csv", sep=",", header = T,skip = 2)
shapeD_tmp <- read.csv2("./data/cell_shape.csv", sep=",", header = T,skip = 2)


## Group them at species level
sporeD <-  sporeD_tmp %>% 
  group_by(species) %>% 
  summarise(sporeforming = Mode(Ability.of.spore.formation)) %>% 
  ungroup()

pathoD <-  pathoD_tmp %>% 
  group_by(species) %>% 
  summarise(pathogen_human = Mode(Pathogenicity..human.)) %>% 
  ungroup()

animal_pathoD <-  animal_pathoD_tmp %>% 
  group_by(species) %>% 
  summarise(pathogen_animal = Mode(Pathogenicity..animal.)) %>% 
  ungroup()

multicelD <-  multicelD_tmp %>% 
  group_by(species) %>% 
  summarise(multicellular_ability = Mode(Multicellular.complex.forming.ability)) %>% 
  ungroup()

shapeD <-  shapeD_tmp %>% 
  group_by(species) %>% 
  summarise(cell.shape = Mode(Cell.shape)) %>% 
  ungroup()

```


# Now do the samefor cell length and cell
They include intervals, so speciel care needs to be take when parsing
```{r}
# Read in the files
widthD_tmp <- read.csv2("./data/cell_width.csv", sep=",", header = T,skip = 2)
lengthD_tmp <- read.csv2("./data/cell_length.csv", sep=",", header = T,skip = 2)

# Function to take the mean of an interval
mean_of_interval <- function(interval){
  split <- str_split(interval,"-")[[1]]
  take_mean <- c(as.double(split[1]), as.double(split[2]))
  mean(take_mean)
}

# Cell length
# Find the mean of the intervals
lengthD_tmp2 <- lengthD_tmp %>% 
  mutate(
    cell.length = 
      ifelse(str_detect(Cell.length,pattern = "-"), mean_of_interval(Cell.length), Cell.length)
    ) %>% 
  mutate(cell.length = as.double(cell.length)) 

# Group by species and find mean
lengthD <- lengthD_tmp2 %>% 
  group_by(species) %>% 
  summarise(cell.length = mean(cell.length)) %>% 
  ungroup()

# Cell width
# Find the mean of the intervals
widthD_tmp2 <- widthD_tmp %>% 
  mutate(
    cell.width = 
      ifelse(str_detect(Cell.width,pattern = "-"), mean_of_interval(Cell.width), Cell.width)
    ) %>% 
  mutate(cell.width = as.double(cell.width)) 

# Group by species and find mean
widthD <- widthD_tmp2 %>% 
  group_by(species) %>% 
  summarise(cell.width = mean(cell.width)) %>% 
  ungroup()


```




### Join them all together and save as Rdata
```{r}
D <- D_tmp %>% 
  left_join(sporeD, by="species") %>% 
  left_join(pathoD, by="species") %>% 
  left_join(animal_pathoD, by="species") %>% 
  left_join(growthD, by="species") %>% 
  left_join(widthD, by="species") %>%  
  left_join(lengthD, by="species") %>% 
  left_join(shapeD, by="species") %>% 
  left_join(multicelD, by="species") %>% 
  left_join(Dar50, by="species") 

# Set the relevant datatypes for all
D <- D %>% 
  mutate(
    species = as.character(species),
    growth_temp = as.character(growth),
    GC.content = as.numeric(as.character(GC.content)),
    sporeforming = as.character(sporeforming),
    pathogen_human = as.character(pathogen_human),
    pathogen_animal = as.character(pathogen_animal),
    multicellular_ability = as.character(multicellular_ability)
  )
D <- D %>% 
  mutate(pathogen_any = ifelse(pathogen_human == 1 | pathogen_animal == 1, factor(1), factor(0))) %>% 
  mutate(pathogen_any = factor(pathogen_any))

# Read to R.data
saveRDS(D, file = "./data/bacdive_growth_ribdif.rds")
```
